GUI/requirements.txt
pyserial
argparse
pandas
numpy
